got a basic FPS counter running as the window title

+++ nl_lib/private/win32/nl_deltatime_win32.c
@@ -12,7 +12,7 @@ internal_function double get_system_time()
     QueryPerformanceFrequency( (LARGE_INTEGER*)&freq);
     QueryPerformanceCounter( (LARGE_INTEGER*)&time);
 
-    double time_seconds = (double)time / (double)freq;
+    double time_seconds = (double)time / freq;
 
     return time_seconds;
 }
